s3tool: GCS-native upload checkpoint (part of LB-1058)
This commit adds a working version of GCS-native upload. Some notes:
* It uses the GCS JSON API
* It supports single-part resumable uploads (copes with IOExceptions too)
* It keeps current implementation's async-logic, client-side encryption &
retry-logic

We tried to add the upload command in a backwards-compatible way and
with changing as less code as possible (for reduced debugging etc).
In particular, this work doesn't try to make s3tool more modular or
more generic.