Added support for loading Google credentials from a file directly
Until now, the Google credentials for the GCS JSON API would be loaded from a
file pointed by the GOOGLE_APPLICATION_CREDENTIALS env var. We added support
for loading them through a file directly. This way we can specify the Google
credentials file in the lb-web config files in an easier way.