Made download command/methods URI handling more consistent
At the CLI level, the download command now expects either a fully qualified
file URI or a prefix + --recursive. A directory-like URI (i.e. ending with ‘/’)
is considered a prefix as well and, hence, the “--recursive” option will be
required.

This change is reflected to the API as well. The `download` method expects a
fully qualified file URI, while the `downloadRecursively` expects a prefix and
it's going to download all files that start with that prefix.

The `downloadDirectory` method was dropped in favor of the
`downloadRecursively`.