Made copy command/methods URI handling more consistent
At the CLI level, the copy command now expects either a fully qualified
file URI or a prefix + --recursive. A directory-like source URI (i.e. ending
with ‘/’) is considered a prefix as well and, hence, the "--recursive" option
will be required. Also, when "--recursive" is used, a directory-like
destination URI is required as well.

This change is reflected to the API as well. The `copy` method expects a
fully qualified file URI, while the `copyRecursively` expects a prefix and
it's going to copy all files starting with that prefix.

The `copyDirectory` method was dropped in favor of the
`copyRecursively`.